Reorganize the global styles sidebar code base
