+++ packages/edit-site/src/components/global-styles/style.scss
@@ -30,6 +30,6 @@
 	}
 }
 
-.edit-site-global-styles-screen-header__description {
+.edit-site-global-styles-header__description {
 	padding: 0 $grid-unit-20;
 }
